+++ src/App.tsx
@@ -5,8 +5,8 @@ import { useDispatch } from 'react-redux';
 
 import AppRouter from './AppRouter';
 import Toast from './Components/Communication/Toast';
-import Header from './Pages/Header';
-import UserDisplay from './Pages/Profile/UserDisplay';
+import Header from './Pages/Sections/Header';
+import UserDisplay from './Pages/Sections/UserDisplay';
 import { getUserWithStoredToken } from './store/user/actions';
 import IceBlueGold from './ui/theme';
