+++ src/Pages/Scores/TotalToto/index.tsx
@@ -13,7 +13,7 @@ import { fetchScoresTotalToto } from '../../../store/scores/actions';
 import { selectTotalToto } from '../../../store/scores/selectors';
 import { UserWithScore } from '../../../store/scores/types';
 import { selectToken } from '../../../store/user/selectors';
-import ScoresBarChart from '../../Charts/ScoresBarChart';
+import ScoresBarChart from '../../Sections/Charts/ScoresBarChart';
 import BreadCrumbsSection from './BreadCrumbsSection';
 
 const TotalToto: React.FC = (): ReactElement => {
